Refactor language model configuration handling: replace direct service method calls with utility function for improved clarity and maintainability
