+++ src/vs/workbench/contrib/chat/common/languageModels.ts
@@ -400,12 +400,6 @@ export interface ILanguageModelsService {
 	 */
 	getModelConfigurationActions(modelId: string): IAction[];
 
-	/**
-	 * Returns a display string for navigation-group configuration properties (e.g., "High" for thinking effort).
-	 * Used by the model picker to show the current config value inline with the model name.
-	 */
-	getModelConfigurationDescription(modelId: string): string | undefined;
-
 	addLanguageModelsProviderGroup(name: string, vendorId: string, configuration: IStringDictionary<unknown> | undefined): Promise<void>;
 
 	removeLanguageModelsProviderGroup(vendorId: string, providerGroupName: string): Promise<void>;
@@ -1182,32 +1176,6 @@ export class LanguageModelsService implements ILanguageModelsService {
 		}
 
 		return actions;
-	}
-
-	getModelConfigurationDescription(modelId: string): string | undefined {
-		const metadata = this._modelCache.get(modelId);
-		const schema = metadata?.configurationSchema;
-		if (!schema?.properties) {
-			return undefined;
-		}
-
-		const currentConfig = this._modelConfigurations.get(modelId) ?? {};
-		const parts: string[] = [];
-
-		for (const [key, propSchema] of Object.entries(schema.properties)) {
-			if (typeof propSchema === 'boolean' || propSchema.group !== 'navigation') {
-				continue;
-			}
-			const value = currentConfig[key] ?? propSchema.default;
-			if (value !== undefined) {
-				// Use enumItemLabels if available
-				const enumIndex = propSchema.enum?.indexOf(value) ?? -1;
-				const displayValue = (enumIndex >= 0 && propSchema.enumItemLabels?.[enumIndex]) ?? String(value);
-				parts.push(displayValue);
-			}
-		}
-
-		return parts.length > 0 ? parts.join(' · ') : undefined;
 	}
 
 	async configureLanguageModelsProviderGroup(vendorId: string, providerGroupName?: string): Promise<void> {
