+++ src/vs/workbench/contrib/chat/test/browser/chatManagement/chatModelsViewModel.test.ts
@@ -135,10 +135,6 @@ class MockLanguageModelsService implements ILanguageModelsService {
 		return [];
 	}
 
-	getModelConfigurationDescription(_modelId: string): string | undefined {
-		return undefined;
-	}
-
 	async configureLanguageModelsProviderGroup(vendorId: string, name?: string): Promise<void> {
 	}
